requirements.txt
Django==3.2.8
psycopg2-binary==2.9.1
djangorestframework==3.12.4
PyJWT==2.3.0
sqlparse==0.4.2